Migrated posts to new framework and improved PostEditorPage
- Ported posts from old framework to new implementation
- Enhanced Posts editor with SettingsMenu
