+++ e2e/tests/admin/posts/post-preview.test.ts
@@ -1,6 +1,6 @@
-import {PostEditorPage} from '../../helpers/pages/admin';
-import {PostFactory, createPostFactory} from '../../data-factory';
-import {expect, test} from '../../helpers/playwright';
+import {PostEditorPage} from '../../../helpers/pages/admin';
+import {PostFactory, createPostFactory} from '../../../data-factory';
+import {expect, test} from '../../../helpers/playwright';
 
 test.describe('Post Preview Modal', () => {
     let postFactory: PostFactory;
